[STARKs] Better code organization (#229)
* [WIP] Deep FRI

* Finish implementation

* Document code

* Fix vulnerability tests and add more

* Define basic traits and structs for Air

* Write basic skeleton of ConstraintEvaluator

* Save evaluator progress

* Complete evaluator basic methods

* Fix some compilation errors

* Finish method for computing composition poly from Evaluation Table

* More fixes

* More progress

* Comment out tests

* [WIP] Implement verifier

* Debugging stuff

* Remove debugging stuff as the boundary ood evaluation was correct

* Uncomment important line

* Start compute_deep_composition_poly refactor

* Finish compute_deep_composition_poly refactor

* Refactor trace function to take the length as a parameter

* Fix hardcoded boundary poly degree on verifier

* Some fixes

* Fix transition_divisors implementation

* Fix transition_divisor once more

* Fix calculation of trace primitive root on the verifier

* Test finally working

* Remove unusued num_assertions field

* [WIP] fiat-shamir integration

* Comment transcript.append line in fri function

* Fix fiat-shamir integration bugs

* Fix boundaryconstraints zerofier test

* Fix clippy suggestiojns

* Remove outdated comments

* Rename compute_transition_evaluations

* Add Fibonacci AIR test for a 17 finite field

* Save refactor WIP

* Refactor compute_deep_composition_poly to generalize for a trace with more than one column

* Add comments to dummy transcript challenge calls in verifier

* Use a better variable name

* Remove debug comment

* Add comments and documentation

* Fix typo

* first steps for multiple poly in trace

* generalization of boundary polys

* Finish refactor of evaluator

* fixing compilation problems

* Refactor evaluation frame module to use multiple trace colums

* Refactor verifier

* Finish first refactor iteration

* Fix overflow in test

* Fix frame bug

* Refactor tables from trace table and frame

* Finish refactor

* Remove unused code

* Remove more unused code

* clippy

* Fix broken code after merge

* Make clippy suggestions

* Remove unused code and fix documentation

* Remove more unused code

* Fix some degree calculation bugs

* Fix degree calculation bugs

* Fix bug for 8 column trace

* test_prove_fib_2_cols first attemp

* Generalize sampling of alphas and betas with Fiat-Shamir for multiple columns

* implementing Debug trait in structs

* Minor fix

* Fix trace table cols() method bug

* Fix transition_divisors bugs - tests are passing

* code cleanup

* iterator improved

* iterators in evaluator

* frame::read_from_trace

* clippy

* refactor in trace::cols

* another refactor in trace cols

* transition_divisors implemented in AIR trait

* refactors in prover and verifier

* query list refactor in prove

* Refactor values function from boundary constraints

* Remove dead code lint

* Remove unused imports

* Remove commented line

* Move integration tests to their own directory

* Fix Fibonacci 17 test

---------
